pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-toolkit"
version = "0.1.2"
description = "Local, lightweight PDF tools (render, split, rotate, page-images)."
requires-python = ">=3.11"
dependencies = [
  "pymupdf>=1.23.0",
  "Pillow>=10.0.0",
  "PyYAML>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
